estandarizado de código respecto a resto de Microservicios, agregado de anotaciones swagger, cambio a uso de DTOs

+++ tp4/scooter-service/src/main/java/com/exa/scooterservice/repository/ScooterRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ScooterRepository extends JpaRepository<Scooter, Long> {
+public interface ScooterRepository extends JpaRepository<Scooter, Integer> {
     List<Scooter> findByState(ScooterState state);
 }
